+++ nengo/tests/test_probe.py
@@ -1,7 +1,6 @@
 import logging
 
 import numpy as np
-import pytest
 
 import nengo
 from nengo.utils.stdlib import Timer
@@ -30,7 +29,6 @@ def test_multirun(Simulator, rng, allclose):
             assert allclose(sim_t[-1], t_sum, rtol=rtol)
 
 
-@pytest.mark.slow
 def test_dts(Simulator, seed, rng):
     """Test probes with different dts and runtimes"""