+++ CHANGES.rst
@@ -53,6 +53,7 @@ Release history
 - Added the ``RLS`` (recursive least-squares) learning rule, which is an online
   version of the least-squares method typically used for offline decoder-solving.
   (`#1611`_, `example <learn-product_>`__)
+- Added the ``SimProbe`` operator, which marks a signal as being probed. (`#1653`_)
 
 **Changed**
 
@@ -125,6 +126,7 @@ Release history
 .. _#1629: https://github.com/nengo/nengo/pull/1629
 .. _#1630: https://github.com/nengo/nengo/pull/1630
 .. _#1640: https://github.com/nengo/nengo/pull/1640
+.. _#1653: https://github.com/nengo/nengo/pull/1653
 .. _learn-product: https://www.nengo.ai/nengo/examples/learning/learn-product.html
 
 3.0.0 (November 18, 2019)